Add default broker in case the broker is not installed.
