+++ src/client/Microsoft.Identity.Client/Platforms/Android/AndroidPlatformProxy.cs
@@ -30,7 +30,6 @@ namespace Microsoft.Identity.Client.Platforms.Android
     {
         internal const string AndroidDefaultRedirectUriTemplate = "msal{0}://auth";
         private const string ChromePackage = "com.android.chrome";
-        private const AndroidBrokerOrchestrator _androidBrokerOrchestrator = new AndroidBrokerOrchestrator();
         // this is used to check if anything can open custom tabs.
         // Must use the classic support. Leaving the reference androidx intent
         //#if __ANDROID_29__
